Make form into its own component
